requirements.txt
# Dependências principais do projeto Flask

# Framework web Flask
Flask==2.3.3

# Carregamento de variáveis de ambiente
python-dotenv==1.0.0

# Dependências do Flask (instaladas automaticamente com Flask)
# Werkzeug==2.3.7
# Jinja2==3.1.2
# MarkupSafe==2.1.3
# click==8.1.7
# blinker==1.6.3
# itsdangerous==2.1.2